rename the namespace to Embeddings
