+++ src/WebJobs.Extensions.OpenAI/Search/ISearchProvider.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.Azure.WebJobs.Extensions.OpenAI.Embedding;
+using Microsoft.Azure.WebJobs.Extensions.OpenAI.Embeddings;
 
 namespace Microsoft.Azure.WebJobs.Extensions.OpenAI.Search;
